feat(beats): add transmission of all beats to separate backoffice socket
