+++ src/modules/beats/artificial-beat-generator.ts
@@ -1,4 +1,4 @@
-import { BeatEvent } from '../events/music-emitter-events';
+import { BeatEvent } from '../events';
 import { BeatGenerator } from './beat-generator';
 
 export interface ArtificialBeatGeneratorParams {
